#156 Removed old voting page
