+++ src/router/index.js
@@ -10,7 +10,7 @@ const LandingPage = () => import('@/views/LandingPage')
 const ImprintPage = () => import('@/views/ImprintPage')
 const TeamPage = () => import('@/views/TeamPage')
 const HowToPlayPage = () => import('@/views/HowToPlayPage')
-const VotingPage = () => import('@/views/NewVotingPage')
+const VotingPage = () => import('@/views/VotingPage')
 const UserView = () => import('@/views/UserView')
 const SteakDrop = () => import('@/views/SteakDropPage')
 const GameOfChains = () => import('@/views/GameOfChainsPage')
