+++ src/views/VotingPage.vue
@@ -1,214 +1,70 @@
 <template>
-  <div>
-    <h2 class="header__h2">
-      Card Voting
-    </h2>
-    <p class="header__p">
-      Take part in the action and share your opinions about community cards. <br>
+  <div align="center">
+    <h1>Voting</h1>
+    <p>
+      Vote to develop the game and earn Tokens with it!<br>
       See results of <router-link :to="{ name: 'VotingResults' }">
         last weeks voting here
       </router-link>.
     </p>
     <br>
-    <div class="voter ccbutton">
-      <div
-        v-if="votingActive"
-        class="box"
-      >
-        <div
-          class="gallery__view__card"
-        >
-          <CardComponent
-            :model="currentCard"
-            :image-u-r-l="currentCard.image"
-          />
-        </div>
-
-        <br>
-        <div class="button-container">
-          <button @click="vote('fair_enough')">
-            Fair Enough
-          </button>
-          <button @click="vote('overpowered')">
-            Overpowered
-          </button>
-          <button @click="vote('underpowered')">
-            Underpowered
-          </button>
-          <button @click="vote('inappropriate')">
-            Inappropriate
-          </button>
-        </div>
-      </div>
-      <div v-if="!votingActive">
-        <span>You cannot vote on cards. Please login with your wallet.</span>
-        <div v-if="noMoreVotesLeft">
-          <span>You have already voted on all cards. Come back tomorrow to vote again.</span>
-        </div>
-        <div v-if="unregistered">
-          <span>You are not registered. To vote on cards you have to register, press the Join button to register. </span>
-          <span>Afer registering it might take a few seconds until your account becomes active.</span>
-        </div>
-      </div>
+    <button
+      class="ModeButton"
+      @click="mode = Mode.ENCOUNTER"
+    >
+      <a>Encounters</a>
+    </button>
+    <button
+      class="ModeButton"
+      @click="mode = Mode.COUNCIL"
+    >
+      <a>Council</a>
+    </button>
+    <br>
+    <encounter-voting-component
+      v-if="mode === Mode.ENCOUNTER"
+    />
+    <div v-if="mode === Mode.COUNCIL">
+      <br>
+      <p>Not yet implemented</p>
     </div>
   </div>
 </template>
 
 <script>
-import * as R from 'ramda'
-import CardComponent from '../components/elements/CardComponent'
-import { creditsFromCoins } from '../components/utils/utils.js'
+import EncounterVotingComponent from '@/components/elements/EncounterVotingComponent.vue'
+
+const Mode = {
+  ENCOUNTER: 0,
+  COUNCIL: 1,
+}
 
 export default {
   name: 'VotingPage',
-  components: {CardComponent},
+  components: { EncounterVotingComponent },
   data () {
     return {
-      unregistered: false,
-      votingActive: false,
-      noMoreVotesLeft: false,
-      voteRights: [],
-      cards: [],
-      currentCard: {},
-      config: {
-        minThrowOutDistance: 250,
-        maxThrowOutDistance: 300
-      }
+      mode: Mode.ENCOUNTER,
+      Mode: Mode,
     }
   },
-  watch: {
-    '$store.state.common.wallet.selectedAddress': function () {
-      if (this.$store.getters["getLoggedIn"]) {
-        this.init()
-      }
-    }
-  },
-  mounted () {
-    console.log("loggedin? ", this.$store.getters["getLoggedIn"])
-    if (this.$store.getters["getLoggedIn"]) {
-      this.init()
-    }
-    else {
-      this.notifyInfo('Not logged in', 'You must login to be able to vote.')
-    }
-  },
-  methods: {
-    init () {
-      this.$cardChain.getVotableCards(this.$store.getters['common/wallet/address'])
-        .then(res => {
-          this.$cardChain.getCardList("", "playable", "", "", "", "", "", "")
-          .then(cards => {
-            console.log('getVotableCards:', res)
-            if (res.votables) {
-              this.voteRights = res.votables.voteRights
-
-              var cleaned = []
-              for (var i = 0; i < this.voteRights.length; i++) {
-                if (cards.cardsList.includes(this.voteRights[i].cardId)) {
-                  cleaned.push(this.voteRights[i])
-                }
-              }
-
-              this.voteRights = cleaned
-
-              if (this.voteRights.length > 0) {
-                console.log('voteRights:', this.voteRights)
-
-                this.getNextCard()
-                  .then(() => {
-                    this.showNextCard()
-                  })
-                this.getNextCard()
-              } else {
-                this.votingActive = false
-              }
-            } else if (res.votables === null) {
-              this.votingActive = false
-              this.noMoreVotesLeft = true
-              console.log('no more voting rights')
-            } else if (res.unregistered === true) {
-              this.unregistered = true
-              this.notifyFail('NOT REGISTERED', 'You are not registered in the blockchain. Please register to obtain voting rights.')
-            } else if (res.noVoteRights === true) {
-              this.noMoreVotesLeft = true
-              this.notifyFail('No Vote Rights', 'You do not have any voting rights, therefore you cannot vote on cards.')
-            } else {
-              this.votingActive = false
-              console.error('getVotableCards returned non-readable data: ', res)
-            }
-          })
-        })
-        .catch(res => {
-          this.notifyFail('OH NOES', res)
-        })
-    },
-    vote (type) {
-      this.getNextCard()
-      this.$cardChain.voteCardTx(this.currentCard.id, type)
-        .then(_ => {
-          this.$cardChain.updateUserCredits()
-        })
-      console.log('vote cast for cardid', this.currentCard.id, 'voted: ', type)
-
-      if (R.isEmpty(this.cards)) {
-        if (!R.isEmpty(this.voteRights)) {
-          this.getNextCard()
-            .then(this.showNextCard())
-        } else {
-          this.votingActive = false
-          this.noMoreVotesLeft = true
-        }
-      } else {
-        this.showNextCard()
-      }
-    },
-    getNextCard () {
-      console.log('votingRights.length:', this.voteRights.length)
-      if (this.voteRights.length > 0) {
-        let nextCard = R.last(this.voteRights)
-        console.log("nextCard", nextCard)
-        this.voteRights = R.dropLast(1, this.voteRights)
-
-        return this.$cardChain.getCard(nextCard.cardId)
-          .then(res => {
-            console.log("res", res)
-            let parsedCard = this.$cardChain.cardObjectToWebModel(res)
-            console.log('currentCard', parsedCard)
-            if (parsedCard) {
-              this.cards.push(parsedCard)
-              R.last(this.cards).id = nextCard.cardId
-            } else {
-              console.error('card could not be parsed', res)
-            }
-          })
-      } else {
-        console.error('no cards left')
-      }
-    },
-    showNextCard () {
-      this.votingActive = true
-      if (R.isEmpty(this.cards)) {
-        this.votingActive = false
-      }
-      this.currentCard = R.last(this.cards)
-      this.cards = R.dropLast(1, this.cards)
-    }
-  }
 }
 </script>
 
-<style scoped>
-.box {
-  text-shadow: none;
-  position: relative;
-}
-.voter {
-  min-height: 10vh;
-}
-.gallery__view__card {
-  margin: auto;
-  width: 50%;
-  height: 75vh;
-  max-width:500px;
-}
+<style scoped lang="scss">
+  @import "@/scss/variables";
+
+  .ModeButton {
+    padding: 20px;
+    background-color: transparent;
+    border: 0;
+  }
+
+  .ModeButton:hover {
+    background-color: $main-color-c;
+    a {
+      color: black;
+    }
+  }
+
 </style>
